Implement spike recording from Iowa
The Iowa experiment now signals events to the recorder on all
button presses in the experiment.

I feel like there might be a cleaner way to setup the spike recorder
client for each app. I think their will be a lot of boiler plate
shared with any experiment devoted to managing the recorder. I think
this could be refactored into the interface.

There seems to be a bug on Mac that is causing the event markers txt
file not to be generated when the recording path is set away from the
default. It is causing tests to fail.